requirements.txt
streamlit>=1.50.0
plotly>=6.5.0
pillow>=11.0.0
numpy>=1.23.0